pyproject.toml
[tool.poetry]
name = "englishwords"
version = "0.1.0"
description = ""
authors = ["wynardtage0 <fishface6666@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
